requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
asyncpg==0.30.0
python-multipart==0.0.9
openai>=1.51.0,<2
httpx==0.27.2
numpy==2.1.2
pdfminer.six==20240706
python-docx==1.1.2
chardet==5.2.0
